+upd in habit and to-do services

+++ Task-Master-Project/src/app/services/to-do.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { Category } from '../to-do/models/models';
+import { Category, Task } from '../to-do/models/models';
 
 @Injectable({
   providedIn: 'root'
@@ -10,30 +10,54 @@ export class ToDoService {
   BASE_URL = "http://localhost:8000"
   constructor(private client:HttpClient) { }
 
-  getCategories():Observable<Category[]>{
+  getCategories(user_id : number):Observable<Category[]>{
     return this.client.get<Category[]>(
-      `${this.BASE_URL}/api/categories`
-    )
-  }
-
-  getTasks():Observable<Task[]>{
-    return this.client.get<Task[]>(
-      `${this.BASE_URL}/api/tasks`
+      `${this.BASE_URL}/api/${user_id}/categories`
     )
   }
 
 
-  createCategory(title:string, user : any):Observable<Category>{
-    return this.client.post<Category>(`${this.BASE_URL}/api/categories/`,
+  createCategory(title:string,user_id: number):Observable<Category>{
+    return this.client.post<Category>(`${this.BASE_URL}/api/${user_id}/categories/`,
       {
         title : title,
-        user: user
+        user: user_id
       }
     )
   }
 
-  deleteCategory(category_id: number): Observable<any> {
-    return this.client.delete(`${this.BASE_URL}/api/categories/${category_id}/`)
+  createTask(description:string,user_id: number, category_id : number): Observable<Task>{
+    return this.client.post<Task>(`${this.BASE_URL}/api/${user_id}/tasks/`,
+      {
+        description : description,
+        completed : false,
+        category: category_id
+      }
+    )
+  }
+
+  deleteCategory(category_id: number, user_id: number): Observable<any> {
+    return this.client.delete(`${this.BASE_URL}/api/${user_id}/categories/${category_id}`)
+  }
+
+
+  getTasksByCategory(category_id: number, user_id: number) : Observable<Task[]>{
+    return this.client.get<Task[]>(
+      `${this.BASE_URL}/api/${user_id}/categories/${category_id}/tasks/`
+    )
+  }
+
+  
+  getTasks(user_id : number):Observable<Task[]>{
+    return this.client.get<Task[]>(
+      `${this.BASE_URL}/api/${user_id}/tasks`
+    )
+  }
+
+
+
+  deleteTask(task_id: number, user_id:number): Observable<any> {
+    return this.client.delete(`${this.BASE_URL}/api/${user_id}/tasks/${task_id}`)
   }
 
 }
